% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRM.R
\name{autoplot.SRM}
\alias{autoplot.SRM}
\title{Autoplot a SRM object}
\usage{
\method{autoplot}{SRM}(obj, facet_names = NULL)
}
\arguments{
\item{obj}{SRM object}

\item{facet_names}{Titles of the facets}
}
\value{
A ggplot object
}
\description{
Autoplot a SRM object
}
