% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRMroll.R
\name{autoplot.SRMroll}
\alias{autoplot.SRMroll}
\title{SRMroll autoplot method}
\usage{
\method{autoplot}{SRMroll}(obj, facet_names = NULL)
}
\arguments{
\item{facet_names}{}
}
\description{
SRMroll autoplot method
}
