% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRiskFCeval.R
\name{autoplot.SystemicRiskFCeval}
\alias{autoplot.SystemicRiskFCeval}
\title{Autoplot method for SystemicRiskFCeval}
\usage{
\method{autoplot}{SystemicRiskFCeval}(obj)
}
\arguments{
\item{obj}{SystemicRiskFCeval object}
}
\description{
Autoplot method for SystemicRiskFCeval
}
