% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRM.R
\name{cov.SRM}
\alias{cov.SRM}
\title{SRM p arameter covariance estimation}
\usage{
cov.SRM(SRM_object, method, ...)
}
\arguments{
\item{SRM_object}{A SRM object}

\item{method}{The method to compute the standard errors. Either "asymptotic" or "boot"}

\item{...}{Further input parameters}
}
\value{
covariance matrix
}
\description{
SRM p arameter covariance estimation
}
