% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CoVaR_tdist}
\alias{CoVaR_tdist}
\title{Co Value at Risk of a bivariate t distribution}
\usage{
CoVaR_tdist(alpha, beta, nu, H)
}
\arguments{
\item{alpha}{Probability level for the CoVaR}

\item{beta}{Probability level for the VaR}

\item{nu}{Degrees of freedom of the bivariate t-distribution}

\item{H}{Covariance matrix of the bivariate t-distribution (2x2)}
}
\description{
Co Value at Risk of a bivariate t distribution
}
