% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{diagnose}
\alias{diagnose}
\title{Outputs Effective Sample Size Diagonis for MCMC run}
\usage{
diagnose(res)
}
\arguments{
\item{res}{output from \code{\link{sample_HierarchicalModel}}.}
}
\description{
Computes the Effective Sample Size using the method
\code{effectiveSize} in of the package \code{coda}.
}
\details{
Currently only works with L where the diagonal is 0. The function
ignores the diagonal and tries to determine from the row and column
sums which parts of the matrix are 0.
}
