% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRM.R
\name{forecast.SRM}
\alias{forecast.SRM}
\title{SRM forecast method}
\usage{
\method{forecast}{SRM}(obj, newdata = NULL)
}
\arguments{
\item{obj}{SRM object}

\item{newdata}{new data set to use for forecasting (for models with regressors or in dynamic models if the model is no re-estimated in this specific time step)}
}
\value{
Vector of forecasts
}
\description{
SRM forecast method
}
