% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{genL}
\alias{genL}
\title{Generate Liabilities Matrix from Prior}
\usage{
genL(model)
}
\arguments{
\item{model}{a model for p and lambda.}
}
\value{
A list consisting of a liabilities matrix and the parameter
vector theta.
}
\description{
Generates a libabilities matrix using a the prior distribution from
a given model for p and lambda.
}
\examples{
n <- 5
m <- Model.Indep.p.lambda(Model.p.BetaPrior(n),
                          Model.lambda.GammaPrior(n,scale=1e-1))
genL(m)

}
