% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_FindFeasibleMatrix.R
\name{getfeasibleMatr}
\alias{getfeasibleMatr}
\title{Creates a feasible starting matrix}
\usage{
getfeasibleMatr(L, A)
}
\arguments{
\item{L}{Vector of row sums}

\item{A}{Vector of column sums}
}
\value{
A matrix with nonnegative entries and given row/column sums
and 0 on the diagonal.
}
\description{
Creates a matrix with nonnegative entries, given row and column
sums and 0 on the diagonal. Superseeded by the more flexible findFeasibleMatrix.
}
\examples{
getfeasibleMatr(c(0.5,1,0),c(0.5,0,1))
getfeasibleMatr(rep(1,4),rep(1,4))
getfeasibleMatr(2^(1:3),2^(3:1))
getfeasibleMatr(1:5,1:5)
getfeasibleMatr(1:5,5:1)
}
