% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRiskFCeval.R
\name{loss_CoVaR}
\alias{loss_CoVaR}
\title{CoVaR part of the joint (VaR, CoVaR) loss function}
\usage{
loss_CoVaR(x, y, VaR, CoVaR, alpha, b = 1)
}
\arguments{
\item{x}{Observation (corresponding to the VaR series)}

\item{y}{Observation (corresponding to the CoVaR series)}

\item{VaR}{VaR forecasts}

\item{CoVaR}{CoVaR forecasts}

\item{alpha}{CoVaR quantile level alpha}

\item{b}{degree of homogeneity (b>=0)}
}
\description{
CoVaR part of the joint (VaR, CoVaR) loss function
}
