% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRiskFCeval.R
\name{loss_MES}
\alias{loss_MES}
\title{MES part of the joint (VaR, MES) loss function}
\usage{
loss_MES(x, y, VaR, MES)
}
\arguments{
\item{x}{Observation (corresponding to the VaR series)}

\item{y}{Observation (corresponding to the MES series)}

\item{VaR}{VaR forecasts}

\item{MES}{MES forecasts}
}
\description{
MES part of the joint (VaR, MES) loss function
}
