% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRiskFCeval.R
\name{loss_VaR}
\alias{loss_VaR}
\title{VaR part of the joint (VaR, CoVaR) loss function}
\usage{
loss_VaR(x, VaR, beta, b = 1)
}
\arguments{
\item{x}{Observation}

\item{VaR}{VaR Forecast}

\item{beta}{Quantile level beta}

\item{b}{degree of homogeneity (b>=0)}
}
\description{
VaR part of the joint (VaR, CoVaR) loss function
}
