% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MES_tdist}
\alias{MES_tdist}
\title{Marginal Expected Shortfall of a bivariate t distribution}
\usage{
MES_tdist(beta, nu, H)
}
\arguments{
\item{beta}{Probability level for the VaR}

\item{nu}{Degrees of freedom of the bivariate t-distribution}

\item{H}{Covariance matrix of the bivariate t-distribution (2x2)}
}
\description{
Marginal Expected Shortfall of a bivariate t distribution
}
