% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExponentialFitnessLinkFun.R
\name{Model.fitness.meandegree}
\alias{Model.fitness.meandegree}
\title{Mean out-degree of a random node  the fitness model}
\usage{
Model.fitness.meandegree(alpha, beta, gamma = 1)
}
\arguments{
\item{alpha}{Exponent of the power law of the degree distribution. Must be <0.}

\item{beta}{Lower endpoint of the relative expected out degree
(expected out degree divided by n-1). Must be >=0.}

\item{gamma}{Upper endpoint of the relative expected out degree
(expected out degree divided by n-1). Must be at least beta and at most 1.}
}
\description{
Computes the relative mean out-degree of a randomly chosen node
given fitness \code{x} in the fitness model implemented in
\code{\link{Model.additivelink.exponential.fitness}}. The function
returns the mean out-degree divided by n-1.
}
