% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{Model.Indep.p.lambda}
\alias{Model.Indep.p.lambda}
\title{Combination of Independent Models for p and lambda}
\usage{
Model.Indep.p.lambda(model.p, model.lambda)
}
\arguments{
\item{model.p}{model for p.}

\item{model.lambda}{model for lambda.}
}
\description{
Combination of Independent Models for p and lambda
}
\examples{
n <- 5
m <- Model.Indep.p.lambda(Model.p.BetaPrior(n),
                          Model.lambda.GammaPrior(n,scale=1e-1))
genL(m)

}
