% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{Model.lambda.constant}
\alias{Model.lambda.constant}
\title{Model for a Constant lambda}
\usage{
Model.lambda.constant(lambda, n)
}
\arguments{
\item{lambda}{paramer for the size of the liabilities. Either a matrix of
dimension n or a single numeric value.}

\item{n}{dimension of matrix.}
}
\value{
the resulting model.
}
\description{
This model assumes that the parameter lambda is known.
}
\examples{
m <- Model.lambda.constant(n=5,lambda=0.25)
m$matr(m$rtheta())
lambda<-matrix(c(NA,1,1,1e-4,NA,1e-4,1e4,1e4,NA),nrow=3)
m <- Model.lambda.constant(n=3,lambda=lambda)
m$matr(m$rtheta())
}
