% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{Model.lambda.GammaPrior}
\alias{Model.lambda.GammaPrior}
\title{Model with Gamma Prior on Lambda}
\usage{
Model.lambda.GammaPrior(n, shape = 1, scale = 1)
}
\arguments{
\item{n}{dimension of matrix}

\item{shape}{shape paramer for prior on
\eqn{\theta}{theta}. Default 1.}

\item{scale}{scale paramer for prior on
\eqn{\theta}{theta}. Default 1.}
}
\description{
Assumes that all elements of lambda are equal to a parameter
\eqn{\theta}{theta}, which has a Gamma prior.
}
