% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{Model.lambda.Gammaprior_mult}
\alias{Model.lambda.Gammaprior_mult}
\title{Model Using Multiple Independent Components}
\usage{
Model.lambda.Gammaprior_mult(Ilambda, shape = 1, scale = 1)
}
\arguments{
\item{Ilambda}{matrix consisting of integers that describe which
component of \eqn{theta}{theta} is used for a given position in the
matrix. Must consist of nonnegative integers using all integers in
the range.}

\item{shape}{shape paramer for prior on
\eqn{\theta}{theta}. Default 1.}

\item{scale}{scale paramer for prior on
\eqn{\theta}{theta}. Default 1.}
}
\value{
the resulting model.
}
\description{
Assumes a multivariate hyperparameter \eqn{\theta}{theta} with each
component following an independent Beta distribution. A matrix
indicates which component \eqn{\theta}{theta} is used for what
component of lambda.
}
