% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{Model.p.BetaPrior}
\alias{Model.p.BetaPrior}
\title{Model for a Random One-dimensional p}
\usage{
Model.p.BetaPrior(n, shape1 = 1, shape2 = 1)
}
\arguments{
\item{n}{dimension of matrix.}

\item{shape1}{first parameter of Beta prior. Default 1.}

\item{shape2}{second  parameter of Beta prior. Default 1.}
}
\value{
the resulting model.
}
\description{
Assumes a Beta prior on the one-dimensional link existence
probabilities p. This model has a one-dimensional parameter.
}
\examples{
m <- Model.p.BetaPrior(5)
m$matr(m$rtheta())
}
