% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{Model.p.Betaprior_mult}
\alias{Model.p.Betaprior_mult}
\title{Model Using Multiple Independent Components}
\usage{
Model.p.Betaprior_mult(Ip, shape1 = 1, shape2 = 1)
}
\arguments{
\item{Ip}{matrix consisting of integers that describe which
component of \eqn{theta}{theta} is used for a given position in the
matrix. Must consist of nonnegative integers (0 encoding forced 0s
in the matrix), using all integers in the range.}

\item{shape1}{first parameter of Beta prior on
\eqn{theta}{theta}. Default 1.}

\item{shape2}{second parameter of Beta prior
\eqn{theta}{theta}. Default 1.}
}
\value{
the resulting model.
}
\description{
Assumes a multivariate hyperparameter \eqn{\theta}{theta} with each
component following an independent Beta distribution. A matrix
indicates which component \eqn{\theta}{theta} is used for what
component of p.
}
