% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{Model.p.constant.nonsquare}
\alias{Model.p.constant.nonsquare}
\title{Model for a constant p and Non-Square Matrices}
\usage{
Model.p.constant.nonsquare(nrow, ncol, p)
}
\arguments{
\item{nrow}{number of rows of the matrix.}

\item{ncol}{number of columns of the matrix.}

\item{p}{existence probability of a link.  A single numeric value.}
}
\value{
the resulting model.
}
\description{
This model assumes that the link existence probabilities of the
matrix are known.
}
\examples{
m <- Model.p.constant.nonsquare(5,3,0.25)
m$matr(m$rtheta())

}
