% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModels.R
\name{Model.p.constant}
\alias{Model.p.constant}
\title{Model for a Constant p}
\usage{
Model.p.constant(n, p)
}
\arguments{
\item{n}{dimension of matrix.}

\item{p}{existence probability of a link. Either a matrix of
dimension n or a single numeric value. A single numeric value leads
to a matrix of existence probabilities that has 0 on the diagonal.}
}
\value{
the resulting model.
}
\description{
This model assumes that the link existence probabilities of the
matrix are known.
}
\examples{
m <- Model.p.constant(5,0.25)
m$matr(m$rtheta())

p <- matrix(c(0,0.99,0.99,0.5,0.5,0,0.01,0.01,0),nrow=3)
m <- Model.p.constant(5,p)
m$matr(m$rtheta())
}
