% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{model_fun}
\alias{model_fun}
\title{SRM Model Function}
\usage{
model_fun(
  theta,
  df,
  prob_level,
  model = "joint_linear",
  risk_measure = "CoVaR",
  forecast = FALSE,
  model_type = "joint",
  m1 = NULL
)
}
\arguments{
\item{theta}{Parameter vector theta}

\item{df}{data.frame/tibble containing the collected data}

\item{prob_level}{list of probability levels beta and alpha}

\item{model}{string or list indicating the model, see Details for more}

\item{risk_measure}{The system risk measure; currently only implemented for the "CoVaR"}

\item{forecast}{logical variable whether the forecast should be produced in the call or not}

\item{model_type}{either "joint", "first" or "second" indicating which model (both, VaR, or CoVaR) should be returned}

\item{m1}{Fitted values for the first (VaR) model; required for model_type="second"}
}
\value{
a list with entries m1 and m2 containing the VaR and CoVaR values as entries
}
\description{
Computes the joint VaR/CoVaR model depending on either a cross-sectional, CoCAViaR or user-written model type. See also \link{SRM}
}
\details{
'model_fun' computes the model, 'nabla_fun' its gradient, and 'theta_fun' some generic information about the respective model parameters.
Currently implemented are the following models:
\itemize{
\item "joint_linear": Linear models with covariates z for both, the VaR and CoVaR
\item "CoCAViaR_SAV_diag": Symmetric Absolute Value (SAV) CoCAViaR model with diagonal matrices A and B
\item "CoCAViaR_SAV_fullA": Symmetric Absolute Value (SAV) CoCAViaR model with full matrix A and and diagonal B
\item "CoCAViaR_SAV_full": Symmetric Absolute Value (SAV) CoCAViaR model with full matrix A and and upper triangular matrix B
\item "CoCAViaR_AS_pos": Asymmetric Slope (AS) CoCAViaR model where only the positive losses (negative returns) act as forcing variables
\item "CoCAViaR_AS_signs": Asymmetric Slope (AS) CoCAViaR model where only the positive and negative losses act as different forcing variables
\item "CoCAViaR_AS_mixed": Asymmetric Slope (AS) CoCAViaR model with 'mixed' forcing variables
}

Own model functions can be written as follows:
ToDo: Specify details!
}
