% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{nabla_joint_linear}
\alias{nabla_joint_linear}
\title{Gradient of the joint linear VaR/CoVaR model}
\usage{
nabla_joint_linear(theta, df, prob_level, risk_measure)
}
\arguments{
\item{theta}{Parameter vector}

\item{df}{Data}

\item{prob_level}{List of the probability levels alpha and beta}

\item{risk_measure}{Systemic risk measure, currently only "CoVaR"}
}
\value{
list of model gradients
}
\description{
Gradient of the joint linear VaR/CoVaR model
}
