% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRiskFCeval.R
\name{plot.SystemicRiskFCeval}
\alias{plot.SystemicRiskFCeval}
\title{Plot method for SystemicRiskFCeval}
\usage{
\method{plot}{SystemicRiskFCeval}(obj, ...)
}
\arguments{
\item{...}{}

\item{SystemicRiskFCeval}{object}
}
\description{
Plot method for SystemicRiskFCeval
}
