% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRM.R
\name{print.SRM}
\alias{print.SRM}
\title{print method for the SRM class}
\usage{
\method{print}{SRM}(obj, digits = 4)
}
\arguments{
\item{obj}{SRM object}

\item{digits}{printed digits after the comma}
}
\description{
print method for the SRM class
}
