% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRM.R
\name{print.summary.SRM}
\alias{print.summary.SRM}
\title{print method for the sum.SRM object}
\usage{
\method{print}{summary.SRM}(sum.SRM_object)
}
\arguments{
\item{sum.SRM_object}{A sum.SRM object}
}
\value{
Nothing
}
\description{
print method for the sum.SRM object
}
