% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRiskFCeval.R
\name{print.SystemicRiskFCeval}
\alias{print.SystemicRiskFCeval}
\title{Print method for SystemicRiskFCeval}
\usage{
\method{print}{SystemicRiskFCeval}(obj, digits = 4)
}
\arguments{
\item{obj}{SystemicRiskFCeval object}

\item{digits}{Printed digits after the comma}
}
\description{
Print method for SystemicRiskFCeval
}
