% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRM.R
\name{summary.SRM}
\alias{summary.SRM}
\title{SRM summary method}
\usage{
\method{summary}{SRM}(SRM_object, method = "asymptotic", ...)
}
\arguments{
\item{SRM_object}{A SRM object}

\item{method}{The method to compute the standard errors. Either "asymptotic" or "boot"}

\item{...}{Further input parameters}
}
\value{
A summary.SRM object
}
\description{
SRM summary method
}
