% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRisk-package.R
\docType{package}
\name{SystemicRisk-package}
\alias{SystemicRisk}
\alias{SystemicRisk-package}
\title{SystemicRisk: Joint Dynamic Systemic Risk Models and Evaluation Tools for the (VaR, CoVaR) and (VaR, MES)}
\description{
The package estimates joint dynamic autoregressive or cross-sectional models of the VaR (Value-at-Risk) (i.e., the quantile) and the CoVaR or MES (Marginal Expected Shortfall) of a bivariate vector of financial returns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TimoDimi/CoQR}
}

}
\author{
\strong{Maintainer}: Timo Dimitriadis \email{timo.dimi@gmail.com} (\href{https://orcid.org/0000-0002-8322-360X}{ORCID})

Authors:
\itemize{
  \item Yannick Hoga \email{Yannick.Hoga@vwl.uni-due.de}
}

}
\keyword{internal}
