% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SystemicRiskFCeval.R
\name{SystemicRiskFCeval}
\alias{SystemicRiskFCeval}
\title{Forecast Evaluation for the Systemic Risk Measures CoVaR and MES}
\usage{
SystemicRiskFCeval(
  data = NULL,
  x = NULL,
  y = NULL,
  VaR1 = NULL,
  VaR2 = NULL,
  CoVaR1 = NULL,
  CoVaR2 = NULL,
  MES1 = NULL,
  MES2 = NULL,
  risk_measure = "CoVaR",
  beta = 0.95,
  alpha = 0.95,
  sided = "onehalf",
  cov_method = "HAC",
  sig_level = 0.05,
  b_VaR = 1,
  b_CoVaR = 1,
  b_MES = 1
)
}
\arguments{
\item{data}{A tsibble containing columns for Date, x,y, VaR1, VaR2, risk_measure1 and risk_measure2, where the latter two are either CoVaR or MES forecasts, as specified in the argument 'risk_measure'.}

\item{x}{Observation (corresponding to the VaR series)}

\item{y}{Observation (corresponding to the CoVaR series)}

\item{VaR1}{Baseline VaR Forecasts}

\item{VaR2}{Competitor VaR Forecasts}

\item{CoVaR1}{Baseline CoVaR Forecasts}

\item{CoVaR2}{Competitor CoVaR Forecasts}

\item{MES1}{Baseline MES Forecasts}

\item{MES2}{Competitor MES Forecasts}

\item{risk_measure}{The Systemic Risk Measure to be evaluated; either CoVaR or MES}

\item{beta}{Probability level of the VaR}

\item{alpha}{Probability level of the CoVaR}

\item{sided}{"onehalf" or "two"; see Fissler and Hoga (2021, arXiv) for details}

\item{cov_method}{covariance estimation method; "HAC" or "iid"}

\item{sig_level}{Significance level for the illustrative plots}
}
\description{
Forecast Evaluation for the Systemic Risk Measures CoVaR and MES
}
